@@ -1,17 +1,12 @@
-import {
-  useRouteMatch,
-  Switch,
-  Route
-} from "react-router-dom";
-import * as React from 'react'
+// Not yet implemented in the App.js file
+// To use this functionality, add a Route to this content
+
+import React from 'react'
 import CorsExampleComp from './CorsExampleComp'
 
 const CorsExample = (() => {
-  let match = useRouteMatch();
-  return(
-    <Switch>
-      <Route path={`${match.url}/`} component={CorsExampleComp} />
-    </Switch>
+  return (
+    <CorsExampleComp />
   )
 })
 
