@@ -1,4 +1,7 @@
-import * as React from 'react'
+// Not yet implemented in the App.js file
+// To use this functionality, add a Route to this content
+
+import React from 'react'
 
 import { ComponentsProvider } from '@looker/components'
 
